DEV-4125 extracted classes, applied formatting
