+++ anjlab-tapestry-quartz/src/main/java/com/anjlab/tapestry5/services/quartz/AnyExecutionMatcher.java
@@ -1,0 +1,12 @@
+package com.anjlab.tapestry5.services.quartz;
+
+import org.quartz.JobExecutionContext;
+
+public class AnyExecutionMatcher implements ExecutionMatcher
+{
+    @Override
+    public boolean matched(JobExecutionContext context)
+    {
+        return true;
+    }
+}
