+++ anjlab-tapestry-quartz/src/main/java/com/anjlab/tapestry5/services/quartz/ExecutionMatcher.java
@@ -1,0 +1,8 @@
+package com.anjlab.tapestry5.services.quartz;
+
+import org.quartz.JobExecutionContext;
+
+public interface ExecutionMatcher
+{
+    boolean matched(JobExecutionContext context);
+}
