+++ anjlab-tapestry-quartz/src/main/java/com/anjlab/tapestry5/services/quartz/KeyValueExecutionMatcher.java
@@ -1,0 +1,21 @@
+package com.anjlab.tapestry5.services.quartz;
+
+import org.quartz.JobExecutionContext;
+
+public class KeyValueExecutionMatcher implements ExecutionMatcher
+{
+    private final String keyName;
+    private final String value;
+
+    public KeyValueExecutionMatcher(String keyName, String value)
+    {
+        this.keyName = keyName;
+        this.value = value;
+    }
+
+    @Override
+    public boolean matched(JobExecutionContext context)
+    {
+        return value.equals(context.getTrigger().getJobDataMap().get(keyName));
+    }
+}
