+++ anjlab-tapestry-quartz/src/test/java/com/anjlab/tapestry5/services/quartz/SchedulerTest.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2014 AnjLab
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -32,7 +32,6 @@ import java.util.concurrent.TimeoutException;
 
 import static org.quartz.DateBuilder.futureDate;
 
-
 public class SchedulerTest
 {
     protected static Registry registry;
@@ -41,8 +40,8 @@ public class SchedulerTest
     public static void setup()
     {
         registry = new RegistryBuilder().add(
-                    TapestryModule.class,
-                    QuartzModule.class)
+                TapestryModule.class,
+                QuartzModule.class)
                 .build();
 
         registry.performRegistryStartup();
@@ -51,29 +50,28 @@ public class SchedulerTest
     @Test
     public void testRunJob()
             throws SchedulerException, InterruptedException,
-                   ExecutionException, TimeoutException
+            ExecutionException, TimeoutException
     {
         Scheduler scheduler = registry.getService(Scheduler.class);
-        
+
         JobDetail job = JobBuilder.newJob(HelloJob.class).build();
-        
+
         job.getJobDataMap().put("username", "John Smith");
-        
+
         Trigger trigger = TriggerBuilder.newTrigger().startNow().build();
-        
+
         QuartzJobFuture<String> future =
                 new QuartzJobFuture<>(scheduler, job.getKey());
-        
+
         scheduler.scheduleJob(job, trigger);
-        
+
         String result = future.get(5, TimeUnit.SECONDS);
-        
+
         String appVersion = registry.getService(SymbolSource.class)
                 .valueForSymbol(SymbolConstants.APPLICATION_VERSION);
-        
+
         Assert.assertEquals("username = John Smith, app version = " + appVersion, result);
     }
-
 
     @Test
     public void testQuartzFutureMatchByExecutionId()
@@ -82,7 +80,7 @@ public class SchedulerTest
     {
         Scheduler scheduler = registry.getService(Scheduler.class);
 
-        JobDetail job = JobBuilder.newJob(HelloJob.class).withIdentity("job1","group1").build();
+        JobDetail job = JobBuilder.newJob(HelloJob.class).withIdentity("job1", "group1").build();
         Trigger trigger = TriggerBuilder
                 .newTrigger()
                 .startAt(futureDate(10, DateBuilder.IntervalUnit.MINUTE))
@@ -94,7 +92,6 @@ public class SchedulerTest
         final String executionId = UUID.randomUUID().toString();
         jobDataMap.put("executionId", executionId);
         jobDataMap.put("username", "John Smith");
-
 
         QuartzJobFuture<String> future =
                 new QuartzJobFuture<>(scheduler, job.getKey(), "executionId", executionId);
